+++ crates/gen/src/types/mod.rs
@@ -10,7 +10,6 @@ mod namespace;
 mod object;
 mod param;
 mod required_interface;
-mod required_interfaces;
 mod r#struct;
 mod r#type;
 mod type_guid;
@@ -29,7 +28,6 @@ pub(crate) use r#enum::Enum;
 pub(crate) use r#struct::Struct;
 pub(crate) use r#type::Type;
 pub(crate) use required_interface::*;
-pub(crate) use required_interfaces::*;
 pub(crate) use type_guid::{GuidConstant, TypeGuid};
 pub(crate) use type_kind::TypeKind;
 pub(crate) use type_name::TypeName;
